+++ extensions/vuengine-studio-extension/src/editors/browser/components/FontEditor/Alphabet/Alphabet.tsx
@@ -48,5 +48,5 @@ export default function Alphabet(props: AlphabetProps): JSX.Element {
                 </div>
             ))
         }
-    </div >;
+    </div>;
 }