+++ extensions/vuengine-studio-extension/src/editors/browser/components/MusicEditor/PianoRoll/PianoRollHeader.tsx
@@ -12,5 +12,5 @@ export default function PianoRollHeader(): JSX.Element {
         <div className='pianoRollHeaderAttributes'>
 
         </div>
-    </div >;
+    </div>;
 }