+++ extensions/vuengine-studio-extension/src/editors/browser/components/MusicEditor/Sequencer/Sequencer.tsx
@@ -25,5 +25,5 @@ export default function Sequencer(): JSX.Element {
                 />
             )
         }
-    </div >;
+    </div>;
 }