+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 - [Studio] Added I18n capabilities
 - [Studio] Introduced project files
-- [Studio] Added visual editors for configurations and Specs
 - [Studio] Added Project widget
 - [Studio] Fixed emulator not persisting save
-- [Studio] Added Rumble Pack tool
+- [Studio] Added visual editors and Spec generation capabilities
+- [Studio] Added Rumble Effect, Font, Language editors
 - [Studio] Added automatic port detection for serialport-based flash carts like HyperFlash32 and HyperBoy
 - [Studio] Added auto updater
 - [Studio] Updated to Theia 1.26.0 (from 1.21.0) ([Changelog](https://github.com/eclipse-theia/theia/blob/v1.26.0/CHANGELOG.md))